requirements.txt

pupil-labs-realtime-api==1.1.3
python-dotenv==1.0.0
pyzmq==25.1.1
opencv-python==4.9.0.80
